issue/251 REFACTOR: EditEmail과 EditPassword를 EditEmailOrPassword로 합체
