+++ src/css/EditEmail.css
@@ -5,15 +5,15 @@
   justify-content: center;
 }
 
-.mypage-edit-emailBox {
+.mypage-edit-email-box {
   margin-top: 10rem;
 }
 
-.mypage-edit-email-leftArrow {
+.mypage-edit-leftArrow {
   margin-bottom: 3rem;
 }
 
-.mypage-edit-email-leftArrow button {
+.mypage-edit-leftArrow button {
   background: none;
   border: none;
 }
@@ -23,7 +23,7 @@
   transform: scale(1.1);
 }
 
-.mypage-edit-email-title {
+.mypage-edit-title {
   font-size: 22px;
   font-weight: bold;
   margin-bottom: 4.5rem;
@@ -40,14 +40,14 @@
   text-align: center;
 }
 
-.mypage-edit-email-new_email {
+.mypage-edit-input-box {
   width: 100%;
   display: grid;
   grid-template-columns: 30% 70%;
   margin-bottom: 3rem;
 }
 
-.mypage-edit-email-new_email input {
+.mypage-edit-input-box input {
   width: 100%;
   border: none;
   border-bottom: 1px solid;
@@ -55,28 +55,28 @@
   background: none;
 }
 
-.mypage-edit-email-button {
+.mypage-edit-button {
   display: flex;
   justify-content: flex-end;
 }
 
-.mypage-edit-email-button button {
+.mypage-edit-button button {
   height: 3rem;
   background-color: #d5d5d5;
   border-radius: 10px;
   border: none;
 }
 
-.mypage-edit-email-button button:hover {
+.mypage-edit-button button:hover {
   cursor: pointer;
   transform: scale(1.1);
 }
 
 @media screen and (max-width: 420px) {
-  .mypage-edit-emailBox {
+  .mypage-edit-email-box {
     width: 100%;
   }
-  .mypage-edit-email-title {
+  .mypage-edit-title {
     width: 100%;
     white-space: break-spaces;
   }
