+++ src/css/EditPassword.css
@@ -6,27 +6,14 @@
   justify-content: center;
 }
 
-.mypage-edit-pwBox {
+.mypage-edit-pw-box {
   width: 100%;
   margin-top: 10rem;
-}
-
-.mypage-edit-pw-leftArrow {
-  margin-bottom: 3rem;
-}
-
-.mypage-edit-pw-leftArrow button {
-  background: none;
-  border: none;
 }
 
 .mypage-edit-pw-leftArrow img:hover {
   cursor: pointer;
   transform: scale(1.1);
-}
-
-.mypage-edit-pw-title {
-  margin-bottom: 4.5rem;
 }
 
 .margin-bottom-2rem {
@@ -35,54 +22,6 @@
 
 .inline-block {
   display: inline-block;
-}
-
-.mypage-edit-pw-new_pw {
-  width: 100%;
-  display: grid;
-  grid-template-columns: 30% 70%;
-  margin-bottom: 3.5rem;
-}
-
-.mypage-edit-pw-new_pw input {
-  width: 100%;
-  border: none;
-  border-bottom: 1px solid;
-  text-align: center;
-  background: none;
-}
-
-.mypage-edit-pw-check_pw {
-  width: 100%;
-  display: grid;
-  grid-template-columns: 30% 70%;
-  margin-bottom: 3rem;
-}
-
-.mypage-edit-pw-check_pw input {
-  width: 100%;
-  border: none;
-  border-bottom: 1px solid;
-  text-align: center;
-  background: none;
-}
-
-.mypage-edit-pw-button {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.mypage-edit-pw-button button {
-  height: 3rem;
-  background-color: #d5d5d5;
-  border-radius: 10px;
-  border: none;
-}
-
-.mypage-edit-pw-button button:hover {
-  cursor: pointer;
-  transform: scale(1.1);
 }
 
 @media screen and (max-width: 420px) {
